Refactored network code
Refactored the code so a network module can be set up as either a
server or a client. createNetWorkModule does the common setup and
startServer and startClient are more specific.  The client code
is broken at the moment but the server works with the client from
here:

http://www.linuxhowtos.org/C_C++/socket.htm

Had to copy createMemoryShare from engine.c, this is temporary